Update updater.sh
Script will now allow you to pick where it "works" so where it temp stores the hashes.txt file and the linux64 file also the signing key. The default is set to /tmp this way the script doesn't risk removing important files that have the same name as the files needed when downloading and updating.